Dynamically displaying content on the page.
Using 'v-if' and 'v-show'.
The two act the same at a presentation level.
Behind the scenes, they differ. v-if will completely remove an element if its conditions are not true, whereas v-show will retain the element and just set its 'display' tag to'none'.

v-show should be used if you are loading in computationally expensive elements, as it will retain the element in the front-end, it will just not be on display. v-if on the other hand will load in the element each time.